Revert "refactor: changed "saveables" to "entities""
This reverts commit cd3dad816d04f314cd730a50078fd6d407a545a5.
